# api/openapi.yml
openapi: 3.0.3
info:
  title: Sky Towers API
  version: 1.0.0
  description: Local API of the Sky Towers hotel management system.
paths:
...
  /api/v1/followups/tasks/count:
    get:
      operationId: followups_tasks_count_retrieve
      description: Badge numbers for the login screen, sidebar and top bar (spec §10.4
        ``tasks?status=open&count``).
      tags:
      - followups
      security:
      - tokenAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCount'
          description: ''
...
components:
  schemas:
...
    TaskCount:
      type: object
      properties:
        late:
          type: integer
        today:
          type: integer
        upcoming:
          type: integer
        system:
          type: integer
        count:
          type: integer
          description: 'Open tasks (open, snoozed, waiting, neglected): the badge
            number.'
      required:
      - late
      - today
      - upcoming
      - system
      - count
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"